refactor : 피드백 반영 - 상속을 통한 반복 줄이기
